% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instants.R
\name{ceiling_decade}
\alias{ceiling_decade}
\title{The end (ceiling) of a decade}
\usage{
ceiling_decade(x)
}
\arguments{
\item{x}{a POSIXct, POSIXlt, Date, chron, yearmon, yearqtr, zoo, zooreg,
timeDate, xts, its, ti, jul, timeSeries, or fts object.}
}
\value{
the end of the decade as a Date object
}
\description{
The end (ceiling) of a decade
}
\examples{
ceiling_decade("2017-10-22 15:01:00")
}
