% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instants.R
\name{days_ago}
\alias{days_ago}
\title{The date x days ago}
\usage{
days_ago(days = 0, tzone = "")
}
\arguments{
\item{days}{an integer specifying the number of days to subtract from the
current}

\item{tzone}{a character vector specifying which time zone you would like to
find the previous date of. tzone defaults to the system time zone set on
your computer.}
}
\value{
the date, x days ago
}
\description{
The date x days ago
}
\examples{
days_ago(3)
days_ago(1) == yesterday()
}
