% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instants.R
\name{days_hence}
\alias{days_hence}
\title{The date x days hence}
\usage{
days_hence(days = 0, tzone = "")
}
\arguments{
\item{days}{an integer specifying the number of days to add from the current}

\item{tzone}{a character vector specifying which time zone you would like to
find the previous date of. tzone defaults to the system time zone set on
your computer.}
}
\value{
the date, x days hence
}
\description{
The date x days hence
}
\examples{
days_hence(3)
days_hence(1) == tomorrow()
}
