% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instants.R
\name{decades_ago}
\alias{decades_ago}
\title{The date x decades ago}
\usage{
decades_ago(decades = 0, tzone = "")
}
\arguments{
\item{decades}{an integer specifying the number of decades to subtract
from the current date}

\item{tzone}{a character vector specifying which time zone you would like to
find the previous date of. tzone defaults to the system time zone set on
your computer.}
}
\value{
The date, x decades ago
}
\description{
The date x decades ago
}
\examples{
decades_ago(1)
decades_ago(2000)
}
