% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instants.R
\name{decades_hence}
\alias{decades_hence}
\title{The date x decades hence}
\usage{
decades_hence(decades = 0, tzone = "")
}
\arguments{
\item{decades}{an integer specifying the number of decades to add to the
current}

\item{tzone}{a character vector specifying which time zone you would like to
find the previous date of. tzone defaults to the system time zone set on
your computer.}
}
\value{
The date, x decades from now
}
\description{
The date x decades hence
}
\examples{
decades_hence(1)
decades_hence(2000)
}
