% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\name{first_monday}
\alias{first_monday}
\alias{first_tuesday}
\title{First and last methods for date-time objects}
\usage{
first_monday(x, unit = "month")

first_tuesday(x, unit = "month")
}
\arguments{
\item{x}{a vector of date-time objects}

\item{unit}{a character string specifying a time unit or a multiple of a unit
to be rounded to. Valid base units are month, bimonth, quarter, season,
halfyear and year.}
}
\description{
Finds the first (or last) date that matches the criteria. Useful for finding
the first/last weekday (e.g. first Monday of the month).
}
\examples{
first_monday("2017-10-22 15:01:00")
first_monday("2019-01-01")
first_tuesday("2017-10-22 15:01:00")
first_tuesday("2019-01-01")
}
\seealso{
[base::round(), lubridate::round_date(), lubridate::floor_date(),
  lubridate::ceiling_date()]
First Monday
}
\keyword{chron}
\keyword{manip}
