% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instants.R
\name{hms}
\alias{hms}
\title{Extract time-of-day values}
\usage{
hms(x)
}
\arguments{
\item{x}{a POSIXct, POSIXlt, Date, chron, yearmon, yearqtr, zoo, zooreg,
timeDate, xts, its, ti, jul, timeSeries, or fts object.}
}
\value{
hms object
}
\description{
Extract time-of-day values
}
\examples{
hms("2017-10-22 15:01:00")
}
