% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instants.R
\name{hours_ago}
\alias{hours_ago}
\title{the time x hours ago}
\usage{
hours_ago(hours = 0, tzone = "")
}
\arguments{
\item{hours}{integer number of hours}

\item{tzone}{a character vector specifying which time zone you would like to
find the previous date of. tzone defaults to the system time zone set on
your computer.}
}
\value{
The time, x hours ago
}
\description{
the time x hours ago
}
\examples{
hours_ago(1)
hours_ago(24)
}
