% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instants.R
\name{hours_hence}
\alias{hours_hence}
\title{the time x hours hence}
\usage{
hours_hence(hours = 0, tzone = "")
}
\arguments{
\item{hours}{integer number of hours}

\item{tzone}{a character vector specifying which time zone you would like to
find the previous date of. tzone defaults to the system time zone set on
your computer.}
}
\value{
The datetime, x hours from now
}
\description{
the time x hours hence
}
