% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instants.R
\name{is.weekday}
\alias{is.weekday}
\title{Is x a weekday?}
\usage{
is.weekday(x)
}
\arguments{
\item{x}{a POSIXct, POSIXlt, Date, chron, yearmon, yearqtr, zoo, zooreg,
timeDate, xts, its, ti, jul, timeSeries, or fts object.}
}
\value{
boolean indicating whether x is a weekday
}
\description{
Is x a weekday?
}
\examples{
is.weekend("2017-08-29")  # FALSE
is.weekend("2017-09-02")  # TRUE
}
