% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instants.R
\name{last_month}
\alias{last_month}
\title{The previous month}
\usage{
last_month(tzone = "")
}
\arguments{
\item{tzone}{a character vector specifying which time zone you would like to
find the previous month of. tzone defaults to the system time zone set on
your computer.}
}
\value{
the previous month as a Date object
}
\description{
The previous month
}
