% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instants.R
\name{last_week}
\alias{last_week}
\title{The previous week}
\usage{
last_week(tzone = "")
}
\arguments{
\item{tzone}{a character vector specifying which time zone you would like to
find the previous week of. tzone defaults to the system time zone set on
your computer.}
}
\value{
the previous week as a Date object
}
\description{
The previous week
}
