% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instants.R
\name{minutes_ago}
\alias{minutes_ago}
\title{The time x minutes before now}
\usage{
minutes_ago(minutes = 0, tzone = "")
}
\arguments{
\item{minutes}{integer number of minutes}

\item{tzone}{a character vector specifying which time zone you would like to
find the previous date of. tzone defaults to the system time zone set on
your computer.}
}
\value{
The time, x minutes ago
}
\description{
The time x minutes before now
}
\examples{
minutes_ago(1)
minutes_ago(60)
}
