% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instants.R
\name{minutes_hence}
\alias{minutes_hence}
\title{The time x minutes before now}
\usage{
minutes_hence(minutes = 0, tzone = "")
}
\arguments{
\item{minutes}{integer number of minutes}

\item{tzone}{a character vector specifying which time zone you would like to
find the previous date of. tzone defaults to the system time zone set on
your computer.}
}
\value{
The date time, x minutes from now
}
\description{
The time x minutes before now
}
