% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instants.R
\name{months_hence}
\alias{months_hence}
\title{The date x months hence}
\usage{
months_hence(months = 0, tzone = "")
}
\arguments{
\item{months}{an integer specifying the number of weeks to add to the current}

\item{tzone}{a character vector specifying which time zone you would like to
find the previous date of. tzone defaults to the system time zone set on
your computer.}
}
\value{
the date, x months from now
}
\description{
The date x months hence
}
\examples{
months_hence(1)
months_hence(12)
}
