% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instants.R
\name{next_month}
\alias{next_month}
\title{The next month}
\usage{
next_month(tzone = "")
}
\arguments{
\item{tzone}{a character vector specifying which time zone you would like to
find the next month of. tzone defaults to the system time zone set on your
computer.}
}
\value{
the next month as a Date object
}
\description{
The next month
}
