% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instants.R
\name{next_week}
\alias{next_week}
\title{The next week}
\usage{
next_week(tzone = "")
}
\arguments{
\item{tzone}{a character vector specifying which time zone you would like to
find the next week of. tzone defaults to the system time zone set on your
computer.}
}
\value{
the next month as a Date object
}
\description{
The next week
}
