% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instants.R
\name{seconds_ago}
\alias{seconds_ago}
\title{The time x seconds before now}
\usage{
seconds_ago(seconds = 0, tzone = "")
}
\arguments{
\item{seconds}{integer number of seconds}

\item{tzone}{a character vector specifying which time zone you would like to
find the previous date of. tzone defaults to the system time zone set on
your computer.}
}
\value{
The datetime, x seconds ago
}
\description{
The time x seconds before now
}
