% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instants.R
\name{this_month}
\alias{this_month}
\title{The current month}
\usage{
this_month(tzone = "")
}
\arguments{
\item{tzone}{a character vector specifying which time zone you would like to
find the current month of. tzone defaults to the system time zone set on your
computer.}
}
\value{
the current month as a Date object
}
\description{
The current month
}
