% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instants.R
\name{this_week}
\alias{this_week}
\title{The current week}
\usage{
this_week(tzone = "")
}
\arguments{
\item{tzone}{a character vector specifying which time zone you would like to
find the current week of. tzone defaults to the system time zone set on
your computer.}
}
\value{
the current week as a Date object
}
\description{
The current week
}
