% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instants.R
\name{tomorrow}
\alias{tomorrow}
\title{The next day}
\usage{
tomorrow(tzone = "")
}
\arguments{
\item{tzone}{a character vector specifying which time zone you would like to
find the previous date of. tzone defaults to the system time zone set on your
computer.}
}
\value{
the previous date as a Date object
}
\description{
The next day
}
\examples{
tomorrow()
tomorrow("UTC")
}
