% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instants.R
\name{weeks_ago}
\alias{weeks_ago}
\title{The date x weeks ago}
\usage{
weeks_ago(weeks = 0, tzone = "")
}
\arguments{
\item{weeks}{an integer specifying the number of weeks to subtract from the
current}

\item{tzone}{a character vector specifying which time zone you would like to
find the previous date of. tzone defaults to the system time zone set on
your computer.}
}
\value{
the date, x weeks ago
}
\description{
The date x weeks ago
}
\examples{
weeks_ago(1)
weeks_ago(52)
}
