% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instants.R
\name{weeks_hence}
\alias{weeks_hence}
\title{The date x weeks hence}
\usage{
weeks_hence(weeks = 0, tzone = "")
}
\arguments{
\item{weeks}{an integer specifying the number of weeks to add to the
current}

\item{tzone}{a character vector specifying which time zone you would like to
find the previous date of. tzone defaults to the system time zone set on
your computer.}
}
\value{
the date, x weeks from now
}
\description{
The date x weeks hence
}
