% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instants.R
\name{years_ago}
\alias{years_ago}
\title{The date x years ago}
\usage{
years_ago(years = 0, tzone = "")
}
\arguments{
\item{years}{an integer specifying the number of days to subtract from the
current}

\item{tzone}{a character vector specifying which time zone you would like to
find the previous date of. tzone defaults to the system time zone set on
your computer.}
}
\value{
The date, x years ago
}
\description{
The date x years ago
}
