% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instants.R
\name{years_hence}
\alias{years_hence}
\title{The date x years hence}
\usage{
years_hence(years = 0, tzone = "")
}
\arguments{
\item{years}{an integer specifying the number of days to add to the
current}

\item{tzone}{a character vector specifying which time zone you would like to
find the previous date of. tzone defaults to the system time zone set on
your computer.}
}
\value{
The date, x years form now
}
\description{
The date x years hence
}
\examples{
years_hence(1)
years_hence(2000)
}
